+++ doc/CHANGELOG.md
@@ -9,6 +9,8 @@ Platform-specific changes are prefixed with the platform name, otherwise the cha
 - Enabled hybrid PVR DR/DMA vertex submission in driver + sped up pvr_prim() [FG]
 - Add thread priority boosting system [Paul Cercueil = PC]
 - Add performance monitor API [PC]
+- Add/Fixed stat() implementations for all filesystems [AB]
+- **Dreamcast**: Add network speedtest and pvr palette examples [AB]
 
 ## KallistiOS version 2.1.0
 - Cleaned up generated stubs files on a make clean [Lawrence Sebald == LS]